Add files via upload
just tried a few models that you have worked with in the past. But now, there is a handy library called lazypredict, with which you can train all the models available in scikit-learn (and more, like XGBoost and LightGBM) with a single line of code.
lazypredict supports both classification and regression problems, so I will show a brief intro to both.